Integrating latest P4 qt_tools_staging branch to Stash

+++ src/core/lib/data_model/i_item_role.cpp
@@ -1,9 +1,9 @@
 #include "i_item_role.hpp"
 #include "wg_types/hash_utilities.hpp"
 
-#define X( ROLE )																\
-	const char * ROLE##Role::roleName_ = #ROLE;									\
-	size_t ROLE##Role::roleId_ = HashUtilities::compute( #ROLE );					\
+#define X( ROLE )																					\
+	const char * ROLE##Role::roleName_ = #ROLE;														\
+	unsigned int ROLE##Role::roleId_ = static_cast<unsigned int>(HashUtilities::compute( #ROLE ));	\
 	ItemRole ROLE##Role::role_ = std::make_pair( ROLE##Role::roleName_, ROLE##Role::roleId_ );
 	ITEM_ROLES
 #undef X
